Started deploying site redesign.
Major update. Not all pages have the new design yet, but those which do are all
built from Markdown source, using an HTML template. Note that the `dict`, `klo`
and `qq` have their own grahpic profiles, and will be redesigned last.

`markdown2html-old` is the previous (unupdated) version of the markdown tool,
and is still used for the `klo` subdir. It should be phased out before long.

HTML5 Boilerplate is included (and used for the new design).