Inserted default 'data-lang' in HTML (looks less crappy in old Chrome).
It seems the Swedish/English selection does not work in Chrome 6. It looks as
if changes to an HTML tag attribute does not cause re-applying of stylesheets,
which is just too bad, but shouldn't affect that many people (also, page works
fine, except that one cannot change into the Swedish translation).