Major re-write of "{pIqaD}, And How to Read It".
The article now includes a much expanded section on the sounds of Klingon with
pronunciation examples in German, Swedish and English. There are also new IPA
tables outlining the inventory of sounds.

The section on stress has been restructured, so as to make it easier to read,
understand and memorize. The intent was to make it very fast to look up the
stress pattern of any given word.

There's also a (new) section on Klingon dialects, briefly summarizing
difference in pronunciation and grammar.