Changed site 'motto' to something less inane.
The main page used to have the four ideals of Klingon behavior ("Accuracy,
straightforwardness, aggressiveness, and strength") from Power Klingon listed
at the top of the page.

Replaced this (pretty obscure) reference to something a little bit more helpful
in explaining the site purpose for new visitors. ("– Your guide to the Klingon
language").