Improved form focusing on "Klingon Pocket Dictionary" search page.
Search query input is now focused differently depending on whether page is
invoked with a query or not.

If no query is given, search field is focused on page load.

If a query has been entered (i.e. if search result is being shown) then the
search field is not focused (user is more likely to want to scroll the page
than enter a query). However search field has 'tabindex=1' attribute, so a
single <tab> is enough to focus the search field.