Made it possible to build site without a CSS/Javascript minification.
Previously 'make' would fail to build site unless 'yui-compressor' was
installed to handle CSS and Javascript minification. But it is useful (for
debugging purposes) to be able to build the site, even without
minification.

NOTE: It is not possible (nor should it be) to *publish* the site without
minification.

Also added a 'fetch' target for fetching the currently published version of the
page.