Changed CSS for graphic bullets in navigation list.
The bullets where previously centered vertically, which does not look good when
there are several entries above each other, some of which are line breaked.
When this happened became hard to see which of the items in the list each
bullet was associated with.

This happened e.g. in the Quick References on the main page.