Moved tab row to inside page header.
Temporary solution dictched. The tab row is now where it should be.

The additional improvement of letting the page logo overlap the tab row is left
as an exercise for another day. (This will be tricky, because even if the logo
overlaps, the overlapping part should not be clickable. Instead clicking on a
part of the logo that overlaps a tab should count as a click on the *tab*.) Oh,
well -- another day, another sorrow. ;)