makeGlossary() now internally uses new ZDB 'id:' field.
Using the 'id:' field means that we can reload the dictionary at any time,
and immediately access any new information in it, without having to rebuild
any other data structures. (The only problem we might run into is if a word
dissapears from the dictionary.)

This also means that the glossary data structure has been made far slimmer
(relying on the dictionary for all data except the word count). Probably
doesn't matter to much in memory, but makes the stuff written to
localStorage a bit easier on the eye. (Except, of course, that the data
only contains word counts and 'id:' field values -- which isn't exactly
human-readable.)