Added partial flashcard function to Glossary page.
Flashcard functionality is still not advanced enough to actually be
usable.

However it implements some of the most basic functions, namely viewing
of flashcards (so far only Klingon -> English), and the buttons
"Again" and "Known" work. ("Again" indicates that you don't know the
word, and "Known" that the word should be moved immediately to the
"Known Words" list.)

There's also a (debug) list displaying the current queue of words to
be practiced. (Current word is not part of this queue.)