Minor Makefile update.
Added '.DELETE_ON_ERROR'. This deletes files generated unless the rule
terminated with exit status 0. If this is not used, then the failing
files will not be rebuilt when running make again (since there already
exist an output file with a newer creation date that its source).

This is important in rules with uses '>' to write to its output file
(e.g. the CSS and Javascript minification), since these rules *always*
generate an output file, regardless of whether or not they succeed.