Added clear button for Glossary page.
In the Load/Save tab there's now a 'Clear stored data' button, since
in this beta stage, the page has now and then changed the format of
the data stored in localStorage, and might sometimes not load at all
when this happens. The easiest way out of this is simply to delete all
data (this is unlikely to have affected many people since the page
hasn't been released for general consumption).

This button will be removed in the final official version. (At which
point should no longer be needed.)