Glossary page now has some keybindings.
The practice tab has gottes some rudimentary keybindings which should
speeden up the answering of flashcards. Minor changes in help texts
for these keybindings as well.

These keybindings only work when one of the reply buttons under the
practice tab is in focus. The buttons are automatically focused after
a question is displayed, so this should normally not be something a
user has to consider, but it might cause some confusion if the user
for some reason onfocus the buttons. (Doing it this way makes in
unnecessary to unbind and rebind the keys when the user switches
between tabs.)