Added dictionary search form to main page.
The main page now contains a search form for the Klingon Pocket
Dictionary.

This allows for faster searches for people entering the main page,
as well as an extra boon for Google Chrome users which now can use
the "Press <Tab> to search klingonska.org" feature to search the
dictionary without first having to navigate to the web page.