Glossary page now allows for browser history buttons.
Redesigned page tabs so that the browser "back" and "forward" buttons
work. (This also allow you to bookmark any of Glossary Extractor
subpages, which wasn't possible before.)

This requires your browser to support the 'hashchange' event, meaning
that the page most defininitely will not work with MSIE 7, and Safari
4.0 and older (see also http://caniuse.com/hashchange). This will not be
fixed unless there is a need for it, so please file a bug
(https://github.com/zrajm/klingonska.org/issues) if you have an issue
with this.