Minor improvements of glossary page practice tab.
If a Klingon word has several meanings, the number of meanings that the
user will have to guess are now displayed with the question. E.g. for
{qon}, the question is displayed as "{qon} (verb) (2 meanings)". -- In
the other direction the question remains unchanged "1. record (any kind
of recording or writing) 2. compose (verb)".

Part-of-speech is no longer abbreviated in questions (it was too easy to
miss the information that way).