Added commenting capacity in Markdown metadata.
Introduced colon (without a field name) as a comment character in
Markdown metadata. This was done so that I could use `: -*- html -*-` as
the first line of each Markdown file, so that Emacs understands to
display them in html-mode.

`bin/markdown2html` and `bin/markdown2html-old` were updated. And the
line `: -*- html -*-` was added to the beginning of all Markdown source
files.