Added coinage year to 'tag:' field in dictionary.
Also made 'tag:' a required field (since all words will have a coinage
year).

Coinage year is the year of the first occurrence of a word, in some cases
this will be before the publication of The Klingon Dictionary (if the word
occurred in, for example, "Star Trek: The Motion Picture" the coinage year
will be 1979 -- even though that was before the language as such was
invented).