Removed unneeded quotes on attributes in Markdown.
A whole bunch of attribute values in embedded HTML in the Markdown
sources was quoted. Removed quotes for numbers, and "(single) lowercased
strings belonging to a fixed set of strings" (e.g. 'class' and 'id'
attribute values, but not 'alt' or 'title' attribute values -- even when
those consist of just a single string).